requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.17
PyPDF2==3.0.1
requests==2.32.3
python-dateutil==2.9.0
pydantic==2.10.3
python-dotenv==1.0.1 